backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-multipart==0.0.20
requests==2.32.3
python-dotenv==1.0.0
twilio==8.10.0
opencv-python==4.12.0.88
numpy==2.2.6
ultralytics==8.3.225
